Adding va_emerging_tech project - to consider emerging technologies in the context of virginia
This project is meant to build on the virginia project, therefore a combine_data_files.py script was introduced to enable merging data files. It will be used here to merge the current virginia model with additional emerging technology inputs.

Virginia plotting scripts were moved to the virginia project folder